solutions to day 3
there must be a better way than a for loop… initializing a list of matrix indices to check on the map, and then getting the entries in a vectorized way would be marginally better, but it's the index initilization that i'm having trouble with.

it's not just straight up modular arithmetic to get the map to wrap around because matrix indexing starts at 1, not 0 in R. i'm keeping track of whether a wrap around just happened to know if i should bump the index by 1.